Don't throw if deserialising an object of an unavailable type if it is a value for a field that does not exist on the parent type
This is also for backwards compatibility - if there is a new version of
a type C1 being serialised that has a new property that is of a new type
C2 and this data is deserialised where there is an old version of C1
available that does not have this property at all and does not have a
defintion of C2 then it shouldn't matter. The fields that are present on
the older version of C1 should be populated by deserialisation process
but it shouldn't be a problem if there is data referencing unavailable
types if those unavailable types are only required by properties that do
not exist in the current version of the assembly.