Add support for circular references by "reusing" references when reading (and storing "pointer" to objects to reuse when writing)
When serialising and deserialising, a log of objects encountered so far
will be maintained and any references that appear multiple times in the
source data will only be serialised once and then "Reference IDs" will
be recorded any where else that they appear. When deserialised, each
unique reference will be deserialised only once and any Reference IDs
encountered will be translated into references back to those objects. If
an object graph repeats the same reference multiple times then this
*might* result in a smaller serialisation output and cheaper
serialisation and deserialisation processes. But both processes require
object history logs to be constructed, which will use some more
resources and so this behaviour is opt-in for now.