Optimisations I'm looking into work better with larger data (and more instances of the same sets of types) so adjusting benchmark
Instead of one JSON file, there are ten - for variations on the same
data. Still derived from real world data and not selected to skew the
tests other than allowing for data sets where there are more instances
of the same set of types (and this is what my intended use case might
be; a long running service that sends and receives serialised versions
of the same object models back and forth and so I'm happy to have my
benchmark suite target that sort of scenario!)

For .NET Framework, the serialiser comfortably beats Json.NET and is
comparable with the BinaryFormatter. Protobuf is way ahead. For .NET
Core, it still beats Json.NET but by a smaller margin and it beats
BinaryFormatter by a small margin. Protobuf still way ahead
(serialisation 20x faster and deserialisation 10x faster).