Build fast serialisers for simple types whose properties will never be Object Reference IDs - makes serialisation 1.74x faster on net461
This is a bit of a rough cut to see if it looks promising. It makes a
big difference to .NET Framework but no improvement for .NET Core!
Strange.. need to investigate further.

There are TODOs left in the code that need tidying up and there are more
types that could be handled in this manner to speed things up further.