Add a DateTime method to the IWrite interface
It's a common case and it could be optimised slightly (rather than
letting the serialiser write away every field for DateTime instances)
but also because it means that the SharedGeneratedMemberSetter could use
the DateTime-serialising method and so it could produce specialised
serialisers for more types. It hasn't made much difference to the
performance numbers (which is starting to drive me round the bend!
Diagnostic Tools in VS don't show anything obvious and so I'm guessing
at bottlenecks.. and clearly not guessing very well!)