Add more data to the benchmarks (want to see better perf scores and this serialiser does better when types appear in the data more often!)
Since I would like to use this serialiser for long-running services,
adding more data to the benchmarks arguably moves it closer to how I'd
use it in the real world.. AND the serialiser does better when it sees
instances of the same type repeatedly and so the comparisons to ProtoBuf
(on CLR it's 8.9x / 7.9x to serialise / deserialise slower than ProtoBuf
and on .NET Core it's 6.5x / 8.8x slower).

Note that it seems like SampleData files that I added a couple of days
ago to the Benchmarks weren't added to the project and so weren't set ti
Copy-to-Output (and so clean clones of this repo would have had
Benchmarks results inconsistent with those that I've been recording).