Added more data to benchmarks - now ProtoBuf less than 5x faster to serialise and less than 2x faster to deserialise
It's not intended for this library to have performance as its number one
feature, it just needs to be competitive and so I'm happy with this. I
wanted it to be at least 2x as fast as BinaryFormatter, really, as a
point of reference - the benchmarks show 1.6x faster to serialise on
net461 (and so not quite 2x) but 3.5x faster to deserialise while it's
2.2x faster to serialise and 3.8x faster to deserialise on
netcoreapp2.1, which is great.