Added support for "type converters" to change how particular types are represented when serialised
The immediate use case that I have is that linked lists that have 1000s
of items in would cause stack overflows during serialisation because
each item would add a layer to the call stack. However, since the start,
I've envisaged experimenting with adding support for other serialisation
formats such as JSON and it type converters would be helpful there (it's
common to replace specialised list types with simple arrays, for
example). There are a couple of rough edges here but I wanted to commit
something that works and add the subsequent fixes as separate
changesets.