Add binary reader support for "FieldNamePreLoad" data - declare field names first then only use Name IDs for fields in object data
When compiled 'member setters' are generated, they always write only
Field Name Reference IDs - currently, we only try to generate a member
setter for a type after an instance of that type has been written at
least once to ensure that all fields exist in the serialisation data
with their full names and IDs. However, there might be a way to eke out
more performance by using more compiled member setters, even before the
first instance of a type is written out - in this case, it will be
necessary to write extra header data that declares which fields will be
used so that field-name : name-ID information is available to the
reader.