Add a "SpeedyButLimited" option and support it in the Serialiser and BinarySerialisationWriter classes (no public API yet)
This is to provide an option that favours performance over safety
relating to circular reference detection. All backwards / forwards
compatibility support across different versions of serailisation
entities are still supported with this option. When it's exposed through
the public API, it will likely be through a clearly-named new method
(not just a boolean option that can be set to true).