Work in progress: [SpecialisationsMayBeIgnoredWhenSerialising] attribute to help SpeedyButLimited able to generate more member setters
eg. support Dictionary<,> properties that have this attribute on them
(only for cases where you can confident that the property will always be
precisely that type and not an instance of a type derived from
Dictionary<,> AND for cases where the dictionaries will never have a
custom equality comparer because the IEqualityComparer<> field will be
set to null) so that a member setter may be generated for the type that
contains that property.

Outstanding work includes unit tests for member-setter-generation of
types where this attribute is used, an analyser to ensure the attribute
is only applied to applicable fields or property (non-sealed,
non-abstract class field / property types), unit tests for the analyser