(4.1.0) Knocked Microsoft.CodeAnalysis.CSharp.Workspaces dependency version down to see if it will prevent analyser warnings on older VS2017
I noticed warnings about Visual Studio not being able to load the
DeprecatedAnalyser in older versions of VS2017 but upgrading to latest
(15.8.8) got rid of them. I don't have any computers with versions older
than 15.8.0 to test on but hopefully this will help people that do!
Downgraded from 2.8.2 to 2.3.0, which is the earliest version that would
work with .NET Standard 1.6