Make ReadNextArray respect ignoreAnyInvalidTypes for unknown element types
This commit makes it so that if `ReadNextArray` is called with `ignoreAnyInvalidTypes: true`, it will now pass that flag into the `IAnalyseTypesForSerialisation.GetType` method which it calls to get the array's element type. This ensures that if we're reading some unknown data that we want to ignore, it won't fail if the element type is not known.

When the element type is not known (and ignoreAnyInvalidTypes is true), ReadNextArray will now avoid building up an actual Array instance with elements, and will instead just return null at the end (once it's exhausted all the unknown content within the array).